allow setting lower target amounts
The ssp will return a bunch of leaves that add up to the target amount.
It will not create 1 leaf for each target amount. Leaf aggregation is
therefore not possible like this. Also it doesn't matter that the target
amounts are lower, because the remainder will be returned by the ssp as
well.