retry transport errors with custom retry channel
After a phone goes to sleep, when the phone wakes up again, the first grpc request after that will fail with a transport error. In order to ensure that no retry logic is needed on every request, a retry channel is introduced that does the retrying for us. This way every grpc request executes reliably in this scenario.