gracefully handle already claimed transfers
A common case is that a transfer has already been claimed by another
instance of the sdk, or even the same instance, in another flow.
Currently the claim_transfer function then errors. This can leave
payments in a pending state until the next sync round.

In order to fix this, we'll gracefully handle already claimed incoming
transfers. If there is an error from the operator, we'll check what the
state of the transfer is on the operator, by querying the transfer and
determining its state. If the state is already past the expected state,
we'll return OK. The TransferAlreadyClaimed error is now handled inside
claim_transfer, to minimize the amount of code paths that have to handle
this error.